requirements/base.txt
# Base Requirements

Django==1.8.6
bootstrap-admin==0.3.6
django-bootstrap==0.2.4
django-bootstrap-form==3.2
django-bootstrap3==6.2.2
django-compat==1.0.7
django-formtools==1.0
django-transmeta==0.7.3
django-user-accounts==1.3.1
django-widget-tweaks==1.4.1
metron==1.3.5
pyparsing==2.0.3
PyYAML==3.11
six==1.10.0
textile==2.2.2
wheel==0.24.0
XlsxWriter==0.7.7

# Load github requirements
-r github.txt
